Major reorganization of python tests
