+++ treerec/tests/init_marked_mutations.slim
@@ -1,0 +1,6 @@
+defineConstant("n_marks", 20);   // number of marked genomes - also tags a marked_mutation sim
+
+source("init.slim");
+
+initializeMarkedMutations(n_marks);
+
