--- treerec/tests/marked_mutations_setup.slim
@@ -1,7 +1,0 @@
-source("testing_utils.slim");
-
-// SETUP WHATNOT
-
-defineConstant("n_marks", 20);   // number of marked genomes
-initializeMarkedMutations(n_marks);
-
