--- treerec/tests/run_tests.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-set -u
-
-# Runs tests under testRecipes/
-
-FILES="./testRecipes/test_*.slim"
-
-./do_test.sh $FILES
