+++ treerec/tests/test_recipes/test_002_quick_neutral.slim
@@ -1,0 +1,25 @@
+// quick neutral.slim : ^Total mutation frequencies: (.*)$
+//
+// A test script with a quick, vanilla neutral simulation; this is the SLiMgui default model
+//
+initialize() {
+	// TREE SEQUENCE TEST OUTPUT
+	initializeSLiMOptions(keepPedigrees=T); // necessary for pedigreeID
+	source("init_marked_mutations.slim");
+    initializeTreeSeq();
+
+	initializeGenomicElementType("g1", m1, 1.0);
+	initializeGenomicElement(g1, 0, 9);
+	initializeMutationRate(0.0);
+	initializeRecombinationRate(0.2);	
+}
+1 {
+	sim.addSubpop("p1", 500);
+	
+	// TREE SEQUENCE TEST OUTPUT
+	initializeMarks(n_marks);
+}
+200 late() {
+	// TREE SEQUENCE TEST OUTPUT
+	outputMutationResult();
+}
