+++ treerec/tests/test_recipes/test_005 simple sexual X.slim
@@ -3,8 +3,9 @@
 // A test script for the simple X chromosome case
 //
 initialize() {
-    initializeSLiMOptions(keepPedigrees=T); // necessary for pedigreeID
-    source("testing_utils.slim");
+	initializeSLiMOptions(keepPedigrees=T); // necessary for pedigreeID
+	source("init.slim");
+    initializeTreeSeq();
 	initializeSex('X');
 	initializeMutationRate(1e-6);
 	initializeMutationType("m1", 0.5, "f", 0.0);
@@ -14,12 +15,12 @@ initialize() {
 }
 1 {
 	sim.addSubpop("p1", 100);
-    chooseAncestralSamples(5);
+	addAncestralSamples(5);
 }
 100 {
-    chooseAncestralSamples(5);
+	addAncestralSamples(5);
 }
 
 500 late() { 
-    outputMutationResult();
+	outputMutationResult();
 }
