+++ treerec/tests/test_recipes/test_006 simple sexual Y.slim
@@ -3,8 +3,9 @@
 // A test script for the simple Y chromosome case
 //
 initialize() {
-    initializeSLiMOptions(keepPedigrees=T); // necessary for pedigreeID
-    source("testing_utils.slim");
+	initializeSLiMOptions(keepPedigrees=T); // necessary for pedigreeID
+	source("init.slim");
+    initializeTreeSeq();
 	initializeSex('Y');
 	initializeMutationRate(1e-6);
 	initializeMutationType("m1", 0.5, "f", 0.0);
@@ -14,12 +15,12 @@ initialize() {
 }
 1 {
 	sim.addSubpop("p1", 100);
-    chooseAncestralSamples(5);
+	addAncestralSamples(5);
 }
 100 {
-    chooseAncestralSamples(5);
+	addAncestralSamples(5);
 }
 500 late() { 
-    outputMutationResult();
-    cat("\nTotal mutation frequencies: " + sum(sim.mutationFrequencies(NULL)) + "\n"); 
+	outputMutationResult();
+	cat("\nTotal mutation frequencies: " + sum(sim.mutationFrequencies(NULL)) + "\n"); 
 }
