+++ treerec/tests/test_recipes/test_007 cloning and selfing.slim
@@ -3,8 +3,9 @@
 // A test script for clonign and selfing
 //
 initialize() {
-    initializeSLiMOptions(keepPedigrees=T); // necessary for pedigreeID
-    source("testing_utils.slim");
+	initializeSLiMOptions(keepPedigrees=T); // necessary for pedigreeID
+	source("init.slim");
+    initializeTreeSeq();
 	initializeMutationRate(1e-7);
 	initializeMutationType("m1", 0.5, "f", 0.0);
 	initializeGenomicElementType("g1", m1, 1.0);
@@ -17,13 +18,13 @@ initialize() {
 	p1.setSelfingRate(0.4);
 	p1.setCloningRate(0.4);
 
-    chooseAncestralSamples(5);
+	addAncestralSamples(5);
 }
 100 {
-    chooseAncestralSamples(5);
+	addAncestralSamples(5);
 }
 
 500 late() { 
-    outputMutationResult();
-    cat("\nTotal mutation frequencies: " + sum(sim.mutationFrequencies(NULL)) + "\n"); 
+	outputMutationResult();
+	cat("\nTotal mutation frequencies: " + sum(sim.mutationFrequencies(NULL)) + "\n"); 
 }
