+++ treerec/tests/test_recipes/test_013 gene conversion.slim
@@ -3,8 +3,9 @@
 // A test script with a longer vanilla neutral simulation
 //
 initialize() {
-    initializeSLiMOptions(keepPedigrees=T); // necessary for pedigreeID
-    source("testing_utils.slim");
+	initializeSLiMOptions(keepPedigrees=T); // necessary for pedigreeID
+	source("init.slim");
+    initializeTreeSeq();
 	initializeMutationRate(1e-6);
 	initializeMutationType("m1", 0.5, "f", 0.0);
 	initializeGenomicElementType("g1", m1, 1.0);
@@ -16,8 +17,8 @@ initialize() {
 	sim.addSubpop("p1", 200);
 }
 100 {
-    chooseAncestralSamples(5);
+	addAncestralSamples(5);
 }
 500 late() { 
-    outputMutationResult();
+	outputMutationResult();
 }
