+++ treerec/tests/test_recipes/test_038 pure cloning.slim
@@ -1,10 +1,11 @@
-// pure selfing.slim : ^Total mutation frequencies: (.*)$
+// pure cloning.slim : ^Total mutation frequencies: (.*)$
 //
-// A test script for selfing
+// A test script for cloning
 //
 initialize() {
-    initializeSLiMOptions(keepPedigrees=T); // necessary for pedigreeID
-    source("testing_utils.slim");
+	initializeSLiMOptions(keepPedigrees=T); // necessary for pedigreeID
+	source("init.slim");
+    initializeTreeSeq();
 	initializeMutationRate(1e-7);
 	initializeMutationType("m1", 0.5, "f", 0.0);
 	initializeGenomicElementType("g1", m1, 1.0);
@@ -14,11 +15,11 @@ initialize() {
 1 {
 	sim.addSubpop("p1", 500);
 	
-	p1.setSelfingRate(1.0);
+	p1.setCloningRate(1.0);
 }
 100 {
-    chooseAncestralSamples(5);
+	addAncestralSamples(5);
 }
 500 late() { 
-    outputMutationResult();
+	outputMutationResult();
 }
